feat(validation): implemented request validation [PW-320]

+++ backend/java/src/main/java/com/michaelyi/personalwebsite/post/PostService.java
@@ -1,6 +1,7 @@
 package com.michaelyi.personalwebsite.post;
 
 import com.michaelyi.personalwebsite.s3.S3Service;
+import com.michaelyi.personalwebsite.util.StringUtil;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import software.amazon.awssdk.services.s3.model.NoSuchKeyException;
@@ -24,26 +25,26 @@ public class PostService {
             String text,
             MultipartFile image
     ) {
-        Post post = new Post(text);
+        Post post = PostUtil.validateAndConstructPost(text);
+
+        if (PostUtil.isImageValid(image)) {
+            throw new IllegalArgumentException("Image is invalid");
+        }
 
         try {
             readPost(post.getId());
 
             throw new IllegalArgumentException(
-                    "A post with the same title already exists."
+                    "A post with the same title already exists"
             );
         } catch (NoSuchElementException e) {
-            if (image == null) {
-                throw new IllegalArgumentException("An image is required.");
-            }
-
             dao.createPost(post);
             String id = post.getId();
 
             try {
                 s3Service.putObject(id, image.getBytes());
             } catch (IOException err) {
-                throw new IllegalArgumentException("Image could not be read.");
+                throw new IllegalArgumentException("Image could not be read");
             }
 
             return id;
@@ -52,19 +53,27 @@ public class PostService {
 
     public Post readPost(String id)
             throws NoSuchElementException {
+        if (StringUtil.isStringInvalid(id)) {
+            throw new IllegalArgumentException("Id cannot be empty");
+        }
+
         return dao
                 .readPost(id)
                 .orElseThrow(() ->
-                        new NoSuchElementException("Post not found."));
+                        new NoSuchElementException("Post not found"));
     }
 
     public byte[] readPostImage(String id) {
+        if (StringUtil.isStringInvalid(id)) {
+            throw new IllegalArgumentException("Id cannot be empty");
+        }
+
         readPost(id);
 
         try {
             return s3Service.getObject(id);
         } catch (NoSuchKeyException | NoSuchElementException err) {
-            throw new NoSuchElementException("Post image not found.");
+            throw new NoSuchElementException("Post image not found");
         }
     }
 
@@ -77,8 +86,16 @@ public class PostService {
             String text,
             MultipartFile image
     ) {
+        if (StringUtil.isStringInvalid(id)) {
+            throw new IllegalArgumentException("Id cannot be empty");
+        }
+
         Post post = readPost(id);
-        Post updatedPost = new Post(text);
+        Post updatedPost = PostUtil.validateAndConstructPost(text);
+
+        if (PostUtil.isImageValid(image)) {
+            throw new IllegalArgumentException("Image is invalid");
+        }
 
         post.setTitle(updatedPost.getTitle());
         post.setContent(updatedPost.getContent());
@@ -103,6 +120,10 @@ public class PostService {
 
     public void deletePost(String id)
             throws NoSuchElementException {
+        if (StringUtil.isStringInvalid(id)) {
+            throw new IllegalArgumentException("Id cannot be empty");
+        }
+
         readPost(id);
 
         s3Service.deleteObject(id);
